Use pg_roles instead of pg_authid.
pg_roles is a "public" version of pg_authid, which allows use in
more restricted environments (such as RDS).  The downside is
that password comparison doesn't work, as pg_roles only returns
a dummy `********` password.

The side effect is that password will be "updated" every run, even
if the underlying value isn't actually different.